requirements.txt
# AI Playground - Core Dependencies
# Essential packages for AI chatbot functionality

# Core Flask framework
Flask==3.0.0
Werkzeug==3.0.1
MarkupSafe==2.1.3

# Text processing
Markdown==3.5.1

# Azure AI SDK (core functionality)
azure-ai-inference>=1.0.0b4
azure-core>=1.30.0
azure-identity>=1.15.0

# Azure Key Vault SDK (optional - for production Key Vault integration)
azure-keyvault-secrets>=4.7.0

# Environment configuration
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0

# Production server (optional - for deployment)
gunicorn==21.2.0